refactor(transport): own getBus and buildTransportCtx in transport/bus
Moves the shared bus lifecycle and TransportContext builder from
engine/transport-bus.ts into transport/bus.ts so the ./transport public
subpath surfaces getBus naturally. Re-exports the TransportAdapter,
TransportBus, TransportContext, TransportKind, Snapshot, ActionRequest,
ActionResult, Capability, and TransportEvent types from the same module
so plugin authors can implement custom transports through one import.

Renames __resetTransportBusForTests to _resetTransportBusForTests and
tags it @internal to signal it sits outside the supported plugin API.
